Feature: Themed rounds

+++ js/app/app.js
@@ -5,6 +5,7 @@ import { QuizEngine } from '../quiz/quizEngine.js';
 import { I18n, i18nReady, initI18n } from '../core/i18n.js';
 import { showToast } from '../ui/helpers.js';
 import { View } from '../ui/ui.js';
+import { Game } from './game.js';
 
 const finishQuizLoading = () => {
 	Loader.quizzesReady = true;
@@ -83,11 +84,27 @@ const loadQuizManifest = () => {
 export const App = {
 	bindKeypress() {
 		jQuery(document).on('keyup', (event) => {
+			if (event.target.matches('input, textarea, select')) {
+				return;
+			}
 			if (event.shiftKey && event.code == 'KeyP' && QuizEngine.gameInProgress) {
 				View.togglePointsModal();
 			} else if (event.shiftKey && event.altKey && event.code == 'KeyQ') {
 				if (confirm(I18n.t('confirm.easterEgg'))) {
 					DB.purge();
+				}
+			} else if (QuizEngine.gameInProgress) {
+				const key = event.key.toLowerCase();
+				if ((key === 'a' || key === 'o') && jQuery('#getAnswer').is(':visible')) {
+					Game.questionAnswered();
+				} else if (jQuery('#notAnswered').is(':visible')) {
+					if (key === '0') {
+						Game.answeredIncorrectly();
+					} else if (key === '1' && jQuery('#buttonOne').is(':visible')) {
+						Game.answeredCorrectly(1);
+					} else if (key === '2' && jQuery('#buttonTwo').is(':visible')) {
+						Game.answeredCorrectly(2);
+					}
 				}
 			}
 		});
